// src/models/statistics-model.ts
interface StatisticsModel {
        baseAttack: number;
        baseAttackSpeed: number;
        baseArmor: number;
        baseMagicResistence: number;
        baseHealth: number;
        baseMana: number;
        baseMovementSpeed: number;
        criticalStrikeChance: number;
}

export {
    StatisticsModel
}